fixed /chchests suggestions not working changed mod minecraft version over to 1.20.4. marked as compatible with any 1.20 version. might be buggy.
fused MCUtils and BButils into Utils

more changes regarding chchests as announcing system

changed the onServerJoin and Leave Events with their runnable to a custom class to allow to set permanent tasks

changed the entirety of the UpdateListeners to a extending base class and manager

added a custom Position class which is just a storage for a coordinate. adapted toString and equal to make sense.

This Version can be run as of writing this but is generally not recommended. Test Server only.