naming fix with Mod Menu Screen Creator.
added discord jda library as dependency to versions.toml as well as to the fabric project.

updated LICENSE to include information that use of this project is forbidden while banned and sharing modified versions is forbidden until further notice.

added a dashboard when developing where all essential configs are in 1 section.